pts_project/requirements.txt
# Django REST Framework Project Requirements
# Procurement Tracking System (PTS)

# Core Django
Django>=4.2,<5.0
djangorestframework>=3.14.0

# Database
psycopg2-binary>=2.9.9

# Authentication
djangorestframework-simplejwt>=5.3.0
django-allauth>=0.57.0
dj-rest-auth>=5.0.0

# API Documentation
drf-spectacular>=0.26.0

# CORS
django-cors-headers>=4.3.0

# Environment variables
python-decouple>=3.8

# Filtering and search
django-filter>=23.3

# Email
django-anymail>=10.2

# Celery for background tasks (notifications)
celery>=5.3.0
redis>=5.0.0
django-celery-beat>=2.5.0

# Utilities
python-dateutil>=2.8.2
Pillow>=10.0.0
python-docx>=0.8.11
openpyxl>=3.1.0
xlrd>=2.0.0

# Testing
pytest>=7.4.0
pytest-django>=4.5.0
factory-boy>=3.3.0

# Code quality
flake8>=6.1.0
black>=23.9.0
isort>=5.12.0
